Update upload_training.py
The .replace could not handle km's (as were given by ChatGPT) so i added one. The zones are hardcoded in HR, but when a description does not contain Run/Swim etc. (which ChatGPT did not do always) a zone base on the one in the parameter is used (Pace, HR). it maks the script a bit more flexible.